Added navigation library dependency (#2)
This should save time with any candidate who wants to use compose navigation. This is the max version supported by this project without going through the hassle of upgrading all the other dependencies. This doesn't add any usage to the main activity so we can see whether candidates consider using it to begin with, rather than giving them the answer ahead of time.